feat: 회원 관리 기능 구현
- AdminUserService 구현
- User 엔티티 구조 변경: 관리자용 password 필드 추가 & 관리사 계정 id는 Uuid 미적용

+++ be/src/main/java/com/cryptory/be/admin/controller/AdminUserController.java
@@ -50,7 +50,7 @@ public class AdminUserController {
 
     // 사용자 차단/차단 해제
     @PatchMapping("/{userId}/block")
-    public ResponseEntity<?> blockUser(@PathVariable Integer userId, @RequestBody UserBlockRequestDto requestDto) {
+    public ResponseEntity<?> blockUser(@PathVariable Long userId, @RequestBody UserBlockRequestDto requestDto) {
         try{
             adminUserService.blockUser(userId, requestDto);
             return ResponseEntity.ok().build();
@@ -80,7 +80,7 @@ public class AdminUserController {
 
     // 관리자 차단/차단 해제
     @PatchMapping("/admins/{userId}/block")
-    public ResponseEntity<?> blockAdmin(@PathVariable Integer userId, @RequestBody UserBlockRequestDto requestDto){
+    public ResponseEntity<?> blockAdmin(@PathVariable Long userId, @RequestBody UserBlockRequestDto requestDto){
         try{
             adminUserService.blockAdmin(userId, requestDto);
             return ResponseEntity.ok().build();
